Fix #875: move daemon lifecycle into App
