+++ CMakeLists.txt
@@ -168,6 +168,7 @@ target_compile_options(gpu_upsampler PRIVATE
 add_executable(gpu_upsampler_alsa
     src/alsa_daemon.cpp
     src/core/partition_runtime_utils.cpp
+    src/daemon/app/app.cpp
     src/daemon/app/process_resources.cpp
     src/daemon/audio/crossfeed_manager.cpp
     src/daemon/audio/upsampler_builder.cpp
